+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-07-11 10:34:48
-// Commit: refactor(core): extract reusable atomic widgets from core
+// Updated at: 2026-07-11 12:46:38
+// Commit: fix(core): fix memory leak by cancelling core subscription
 
 class coreComponentState {
-  final String commitRef = '2683';
-  final String lastMessage = 'refactor(core): extract reusable atomic widgets from core';
-  final DateTime timestamp = DateTime.parse('2026-07-11T10:34:48+05:00');
+  final String commitRef = '2690';
+  final String lastMessage = 'fix(core): fix memory leak by cancelling core subscription';
+  final DateTime timestamp = DateTime.parse('2026-07-11T12:46:38+05:00');
 }